+++ R/mapdeck_package.R
@@ -28,7 +28,6 @@ NULL
 #' }
 NULL
 
-
 experimental_layer <- function(layer) {
 	message(layer, " is an experimental layer and the function may change without warning")
 }