+++ inst/htmlwidgets/mapdeck_location.js
@@ -23,7 +23,6 @@ function md_change_location( map_id, map_type, location, zoom, pitch, bearing, d
 	    currentZoom = zoom === null ? window[ map_id + 'map'].viewState.zoom : zoom;
 	  }
 
-
 		window[map_id + 'map'].setProps({
 	    viewState: {
 	      longitude: currentLon,
@@ -35,6 +34,12 @@ function md_change_location( map_id, map_type, location, zoom, pitch, bearing, d
 	      transitionDuration: duration
 	    },
 	  });
+
+
+	  //window[ map_id + 'map' ].setProps({
+	  //	viewState: null
+	  //});
+
   }
 }